backend/requirements.txt
alembic==1.13.0
fastapi
fastapi-limiter==0.1.6
email-validator==2.0.0
python-jose==3.3.0
pillow==9.5.0
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
uvicorn==0.29.0
python-multipart==0.0.9
redis==5.0.1
httpx==0.25.2
celery==5.3.6
sqlalchemy-crud-plus==0.0.2
loguru==0.7.2
wait-for-it==2.2.2
fastapi-pagination==0.12.13
asgiref==3.7.2
msgspec==0.18.5
fastapi_versioning==0.10.0
gunicorn==21.2.0
asyncpg==0.29.0
passlib==1.7.4
psycopg2-binary
fast-captcha==0.2.1
phonenumbers==8.13.27
casbin==1.34.0
casbin-async-sqlalchemy-adapter==1.4.0
psutil==5.9.6
cryptography==41.0.7
user-agents==2.2.0
xdbsearchip==1.0.2
itsdangerous==2.1.2
pytest-asyncio==0.23.8
qrcode[pil]==7.4.2
flower==2.0.1
supervisor==4.2.5
jinja2==3.1.3
sentry-sdk[fastapi]
aiohttp
bs4